Rename workers package to workerpool
